backend/pyproject.toml
[tool.poetry]
name = "github-scraping-backend"
version = "0.1.0"
description = "Backend for Github Scraping Project."
authors = ["aditya-ramesh-mysore <mysore.ad@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
